+++ packages/adapter-auto/index.js
@@ -116,5 +116,12 @@ export default () => ({
 		builder.log.warn(
 			'Could not detect a supported production environment. See https://kit.svelte.dev/docs/adapters to learn how to configure your app to run on the platform of your choosing'
 		);
+	},
+	supports: {
+		read: () => {
+			throw new Error(
+				"The read function imported from $app/server only works in certain environments. Since you're using @sveltejs/adapter-auto, SvelteKit cannot determine whether it will work when your app is deployed. Please replace it with an adapter tailored to your target environment."
+			);
+		}
 	}
 });